+++ farm-animal/src/components/navbar/Home.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import HomeBg from "../HomeBg";
-import '../../styles/navbar_/Home.css';
+import HomeBg from "./HomeBg";
+import "../../styles/navbar_/HomeBg.css";
 
 const Home = () => {
   return (
